requirements.txt
aiml==0.9.1
aiodns==1.1.1
aiohttp==3.4.4
aiohttp-socks==0.2
alembic==1.0.0
altgraph==0.15
amqp==2.3.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
backports.csv==1.0.6
beautifulsoup4==4.6.0
billiard==3.5.0.4
BlinkStick==1.1.8
bokeh==0.13.0
boto==2.49.0
boto3==1.9.39
botocore==1.12.39
bottle==0.12.13
bottle-websocket==0.2.9
breadability==0.1.20
bs4==0.0.1
bz2file==0.98
cairocffi==0.9.0
CairoSVG==2.2.1
cchardet==2.1.1
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.2
chromedriver==2.24.1
click==6.7
clinto==0.2.1
coala-utils==0.5.1
codegen==1.0
colorama==0.3.9
colormap==1.0.1
comtypes==1.1.7
ConfigArgParse==0.13.0
constantly==15.1.0
converter==1.0.0
cryptography==2.4.2
cssselect==1.0.3
cssselect2==0.2.1
cx-Freeze==5.1.1
cycler==0.10.0
Cython==0.29
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
Django==1.11.15
django-autoslug==1.9.3
django-celery-results==1.0.1
dnspython==1.15.0
docopt==0.6.2
docutils==0.14
EbookLib==0.17
editdistance==0.4
Eel==0.9.10
elasticsearch==6.3.1
emoji==0.5.0
et-xmlfile==1.0.1
eventlet==0.24.1
extraction==0.2
fake-useragent==0.1.10
fastnumbers==2.1.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-WTF==0.14.2
fpdf==1.7.2
future==0.16.0
gensim==3.6.0
geographiclib==1.49
geopy==1.16.0
gevent==1.3.4
gevent-websocket==0.10.1
google==2.0.1
greenlet==0.4.13
gunicorn==19.9.0
html2text==2018.1.9
html5lib==1.0.1
HTMLParser==0.0.2
hyperlink==18.0.0
idna==2.7
idna-ssl==1.1.0
image==1.5.24
incremental==17.5.0
inflection==0.3.1
ipython==7.0.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==0.24
jdcal==1.4
jedi==0.13.1
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
kombu==4.2.1
language-check==1.1
lxml==4.2.3
macholib==1.9
Mako==1.0.7
markovify==0.7.1
MarkupSafe==1.0
mathparse==0.1.1
matplotlib==2.2.2
mechanize==0.4.0
money-parser==0.0.1
monotonic==1.5
more-itertools==4.3.0
mpld3==0.3
multidict==4.4.1
nbformat==4.4.0
networkx==2.2
newspaper3k==0.2.8
nibabel==2.3.0
nltk==3.4
notify2==0.3.1
numpy==1.15.4
oauthlib==2.1.0
openpyxl==2.5.4
packaging==17.1
pandas==0.23.4
pandas-datareader==0.7.0
parse==1.8.4
parsedatetime==2.4
parso==0.3.1
patterns==0.3
pdfkit==0.6.1
pdfminer2==20151206
peewee==3.8.0
pefile==2017.11.5
pickleshare==0.7.5
Pillow==5.2.0
pipenv==2018.11.26
pipreqs==0.4.9
plotly==3.3.0
pprint==0.1
progress==1.4
prompt-toolkit==2.0.5
proselint==0.10.2
prov==1.5.0
pycares==2.3.0
pycorenlp==0.3.0
pycparser==2.18
pydotplus==2.0.2
pyee==5.0.0
pygame==1.9.4
Pygments==2.2.0
PyHamcrest==1.9.0
PyInstaller==3.3.1
pymongo==3.7.1
pynsist==2.2
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.5
pypiwin32==223
pyppeteer==0.0.24
Pypubsub==4.0.0
pyquery==1.4.0
pyreadline==2.1
PySocks==1.6.8
pyspellchecker==0.1.3
pytesseract==0.2.5
python-dateutil==2.6.1
python-docx==0.8.10
python-dotenv==0.9.1
python-editor==1.0.3
python-telegram-bot==11.1.0
python-twitter==3.4.2
python-vlc==3.0.102
pytrends==4.4.0
pytz==2018.5
pywin32==224
pywinusb==0.4.2
PyYAML==3.13
Quandl==3.4.4
rake-nltk==1.0.4
rdflib==4.2.2
reportlab==3.5.10
requests==2.20.1
requests-download==0.1.2
requests-file==1.4.3
requests-html==0.9.0
requests-oauthlib==1.0.0
retrying==1.3.3
robot==20071211
s3transfer==0.1.13
schedule==0.5.0
scikit-learn==0.20.0
scipy==1.1.0
scraperwiki==0.5.1
seaborn==0.9.0
selenium==3.13.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
slackclient==1.2.1
smart-open==1.7.1
socialreaper==1.0.3
SQLAlchemy==1.2.10
sumy==0.7.0
telegram==0.0.1
telepot==12.7
textblob==0.15.2
textwrap3==0.9.1
tika==1.18
tinycss2==0.6.1
tinysegmenter==0.3
tldextract==2.2.0
tornado==5.1
tqdm==4.24.0
traitlets==4.3.2
tweepy==3.6.0
-e git+https://github.com/twintproject/twint.git@6871a739c1c37cf9d60e2b2fcb2c77e91099f847#egg=twint
twitter-scraper==0.3.0
twitterscraper==0.9.0
Unidecode==1.0.22
urllib3==1.23
validators==0.12.4
vine==1.1.4
virtualenv==16.1.0
virtualenv-clone==0.4.0
w3lib==1.19.0
Wand==0.4.4
wcwidth==0.1.7
WeasyPrint==44
webencodings==0.5.1
websocket-client==0.51.0
websockets==6.0
Werkzeug==0.14.1
whichcraft==0.4.1
win-cli-launchers==0.1
win10toast==0.9
wooey==0.10.0
wordcloud==1.5.0
wrapt==1.10.11
WTForms==2.2.1
wxPython==4.0.3
xmltodict==0.11.0
yarg==0.1.9
yarl==1.2.6
zope.interface==4.5.0
